+++ src/Router/Types.fs
@@ -30,15 +30,15 @@ type RouterConfiguration = private RouterConfiguration of Map<EventName, StreamN
 
 type RouteEvent<'InputEvent, 'OutputEvent> = ProcessedBy -> TracedEvent<'InputEvent> -> TracedEvent<'OutputEvent> option
 
-type internal RouteEventHandler<'InputEvent, 'OutputEvent> =
+type internal RouteEventHandler<'InputEvent, 'OutputEvent, 'Dependencies> =
     | Simple of RouteEvent<'InputEvent, 'OutputEvent>
-    | WithApplication of (PatternRuntimeParts -> RouteEvent<'InputEvent, 'OutputEvent>)
+    | WithApplication of (PatternRuntimeParts<'Dependencies> -> RouteEvent<'InputEvent, 'OutputEvent>)
 
 type internal RouterParts<'InputEvent, 'OutputEvent, 'Dependencies> = {
     Configuration: Configuration<'InputEvent, 'OutputEvent, 'Dependencies> option
     RouterConfiguration: RouterConfiguration option
     RouteToBrokerList: BrokerList option
-    RouteEvent: RouteEventHandler<'InputEvent, 'OutputEvent> option
+    RouteEvent: RouteEventHandler<'InputEvent, 'OutputEvent, 'Dependencies> option
     FromDomain: OutputFromDomain<'OutputEvent> option
     CreateCustomValues: CreateCustomValues<'InputEvent, 'OutputEvent> option
     GetCommonEvent: GetCommonEvent<'InputEvent, 'OutputEvent> option