+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `Dependencies` to `PatternRuntimeParts`
+- Allow `initialize` to return `Result` and `AsyncResult`
 
 ## 22.0.0 - 2022-08-30
 - [**BC**] Add generic dependencies to runtime parts